2020年11月16日20时58分26秒

+++ 工作数据分析_Web/Login.aspx.cs
@@ -17,13 +17,17 @@ public partial class WebPage_Login : System.Web.UI.Page
 
         //每次登陆的时候删除超过1个小时的临时文件
         string path = Server.MapPath("~") + "TmpDownFile";
-        foreach (string file in Directory.EnumerateFiles(path))
+        if (File.Exists(path))
         {
-            if (File.GetLastWriteTime(file)<DateTime.Now.AddHours(-1))
+            foreach (string file in Directory.EnumerateFiles(path))
             {
-                File.Delete(file);
+                if (File.GetLastWriteTime(file) < DateTime.Now.AddHours(-1))
+                {
+                    File.Delete(file);
+                }
             }
         }
+        
     }
